refactor: holdBlockforRender를 hodBlock으로 노출하게 변경
- UI와 데이터를 확실하게 분리
